files/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
asyncpg==0.29.0
redis==5.0.1
pydantic==2.5.0
aiofiles==23.2.1
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
pytest==7.4.3
httpx==0.25.2
